Don't export VisualStudioDocumentTrackerFactory as a Roslyn language service
