--- src/Razor/src/Microsoft.VisualStudio.Editor.Razor/DefaultVisualStudioDocumentTrackerFactory.cs
@@ -1,70 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using System;
-using System.Diagnostics;
-using Microsoft.AspNetCore.Razor.ProjectEngineHost;
-using Microsoft.CodeAnalysis.Razor;
-using Microsoft.CodeAnalysis.Razor.Workspaces;
-using Microsoft.VisualStudio.Editor.Razor.Settings;
-using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Threading;
-
-namespace Microsoft.VisualStudio.Editor.Razor;
-
-internal class DefaultVisualStudioDocumentTrackerFactory : VisualStudioDocumentTrackerFactory
-{
-    private readonly ProjectSnapshotManagerDispatcher _dispatcher;
-    private readonly JoinableTaskContext _joinableTaskContext;
-    private readonly ITextDocumentFactoryService _textDocumentFactory;
-    private readonly IProjectPathProvider _projectPathProvider;
-    private readonly IImportDocumentManager _importDocumentManager;
-    private readonly IProjectSnapshotManagerAccessor _projectManagerAccessor;
-    private readonly IWorkspaceEditorSettings _workspaceEditorSettings;
-    private readonly IProjectEngineFactoryProvider _projectEngineFactoryProvider;
-
-    public DefaultVisualStudioDocumentTrackerFactory(
-        ProjectSnapshotManagerDispatcher dispatcher,
-        JoinableTaskContext joinableTaskContext,
-        IProjectSnapshotManagerAccessor projectManagerAccessor,
-        IWorkspaceEditorSettings workspaceEditorSettings,
-        IProjectPathProvider projectPathProvider,
-        ITextDocumentFactoryService textDocumentFactory,
-        IImportDocumentManager importDocumentManager,
-        IProjectEngineFactoryProvider projectEngineFactoryProvider)
-    {
-        _dispatcher = dispatcher;
-        _joinableTaskContext = joinableTaskContext;
-        _projectManagerAccessor = projectManagerAccessor;
-        _workspaceEditorSettings = workspaceEditorSettings;
-        _projectPathProvider = projectPathProvider;
-        _textDocumentFactory = textDocumentFactory;
-        _importDocumentManager = importDocumentManager;
-        _projectEngineFactoryProvider = projectEngineFactoryProvider;
-    }
-
-    public override VisualStudioDocumentTracker? Create(ITextBuffer textBuffer)
-    {
-        if (textBuffer is null)
-        {
-            throw new ArgumentNullException(nameof(textBuffer));
-        }
-
-        if (!_textDocumentFactory.TryGetTextDocument(textBuffer, out var textDocument))
-        {
-            Debug.Fail("Text document should be available from the text buffer.");
-            return null;
-        }
-
-        if (!_projectPathProvider.TryGetProjectPath(textBuffer, out var projectPath))
-        {
-            return null;
-        }
-
-        var filePath = textDocument.FilePath;
-        var tracker = new DefaultVisualStudioDocumentTracker(
-            _dispatcher, _joinableTaskContext, filePath, projectPath, _projectManagerAccessor, _workspaceEditorSettings, _projectEngineFactoryProvider, textBuffer, _importDocumentManager);
-
-        return tracker;
-    }
-}
