+++ src/Razor/src/Microsoft.VisualStudio.Editor.Razor/IVisualStudioDocumentTrackerFactory.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis.Host;
+using Microsoft.VisualStudio.Text;
 
 namespace Microsoft.VisualStudio.Editor.Razor;
 
-internal abstract class BraceSmartIndenterFactory : ILanguageService
+internal interface IVisualStudioDocumentTrackerFactory
 {
-    public abstract BraceSmartIndenter Create(IVisualStudioDocumentTracker documentTracker);
+    IVisualStudioDocumentTracker? Create(ITextBuffer textBuffer);
 }
